backend/requirements.txt

Django>=5.0
celery>=5.3
redis>=5.0
psycopg2-binary
playwright
cryptography
python-dotenv
gunicorn
djangorestframework
django-cors-headers